Further cleanup and adjustments to fix portable stories
