+++ code/core/package.json
@@ -209,6 +209,11 @@
       "import": "./dist/test/preview.js",
       "require": "./dist/test/preview.cjs"
     },
+    "./test/spy": {
+      "types": "./dist/test/spy.d.ts",
+      "import": "./dist/test/spy.js",
+      "require": "./dist/test/spy.cjs"
+    },
     "./test": {
       "types": "./dist/test/index.d.ts",
       "import": "./dist/test/index.js",
@@ -325,6 +330,9 @@
       ],
       "test/preview": [
         "./dist/test/preview.d.ts"
+      ],
+      "test/spy": [
+        "./dist/test/spy.d.ts"
       ],
       "test": [
         "./dist/test/index.d.ts"
