+++ code/core/src/actions/index.ts
@@ -1,11 +1,5 @@
-import { definePreview } from 'storybook/internal/preview-api';
-
-import * as addonAnnotations from 'storybook/actions/preview';
-
 export * from './constants';
 export * from './models';
 export * from './runtime';
-
-export default () => definePreview(addonAnnotations);
 
 export type { ActionsParameters } from './types';
