+++ code/core/src/actions/preview.ts
@@ -1,2 +1,12 @@
-export * from './addArgs';
+import { definePreview } from 'storybook/internal/preview-api';
+
+import * as addArgs from './addArgs';
+import * as loaders from './loaders';
+
 export * from './loaders';
+
+export default () =>
+  definePreview({
+    ...addArgs,
+    ...loaders,
+  });
