+++ code/core/src/csf/csf-factories.ts
@@ -11,8 +11,8 @@ import type {
   StoryAnnotations,
 } from 'storybook/internal/types';
 
-import actionAnnotations from 'storybook/actions';
-import testAnnotations from 'storybook/test';
+import actionAnnotations from 'storybook/actions/preview';
+import testAnnotations from 'storybook/test/preview';
 
 export interface Preview<TRenderer extends Renderer = Renderer> {
   readonly _tag: 'Preview';
