+++ code/core/src/test/index.test.ts
@@ -2,7 +2,8 @@
 import { describe, it, test } from 'vitest';
 
 import { action } from 'storybook/actions';
-import { expect, fn, isMockFunction, traverseArgs } from 'storybook/test';
+import { expect, fn, isMockFunction } from 'storybook/test';
+import { traverseArgs } from 'storybook/test/preview';
 
 it('storybook expect and fn can be used in vitest test', () => {
   const spy = fn();
