+++ code/core/src/test/index.ts
@@ -1,15 +1,12 @@
 import type { BoundFunctions } from '@testing-library/dom';
-import { userEvent } from '@testing-library/user-event';
+import type { userEvent } from '@testing-library/user-event';
 
-import type { LoaderFunction } from 'storybook/internal/csf';
 import { instrument } from 'storybook/internal/instrumenter';
-import { definePreview } from 'storybook/internal/preview-api';
 
 import { Assertion } from 'chai';
 
 import { expect as rawExpect } from './expect';
-import { clearAllMocks, fn, isMockFunction, resetAllMocks, restoreAllMocks } from './spy';
-import { type queries, within } from './testing-library';
+import { type queries } from './testing-library';
 
 export * from './spy';
 
@@ -42,79 +39,3 @@ export const { expect } = instrument(
 );
 
 export * from './testing-library';
-
-export const resetAllMocksLoader: LoaderFunction = ({ parameters }) => {
-  if (parameters?.test?.mockReset === true) {
-    resetAllMocks();
-  } else if (parameters?.test?.clearMocks === true) {
-    clearAllMocks();
-  } else if (parameters?.test?.restoreMocks !== false) {
-    restoreAllMocks();
-  }
-};
-
-export const traverseArgs = (value: unknown, depth = 0, key?: string): unknown => {
-  // Make sure to not get in infinite loops with self referencing args
-  if (depth > 5) {
-    return value;
-  }
-
-  if (value == null) {
-    return value;
-  }
-  if (isMockFunction(value)) {
-    // Makes sure we get the arg name in the interactions panel
-    if (key) {
-      value.mockName(key);
-    }
-    return value;
-  }
-
-  // wrap explicit actions in a spy
-  if (
-    typeof value === 'function' &&
-    'isAction' in value &&
-    value.isAction &&
-    !('implicit' in value && value.implicit)
-  ) {
-    const mock = fn(value as any);
-
-    if (key) {
-      mock.mockName(key);
-    }
-    return mock;
-  }
-
-  if (Array.isArray(value)) {
-    depth++;
-    return value.map((item) => traverseArgs(item, depth));
-  }
-
-  if (typeof value === 'object' && value.constructor === Object) {
-    depth++;
-    for (const [k, v] of Object.entries(value)) {
-      if (Object.getOwnPropertyDescriptor(value, k)?.writable) {
-        // We have to mutate the original object for this to survive HMR.
-        (value as Record<string, unknown>)[k] = traverseArgs(v, depth, k);
-      }
-    }
-    return value;
-  }
-  return value;
-};
-
-export const nameSpiesAndWrapActionsInSpies: LoaderFunction = ({ initialArgs }) => {
-  traverseArgs(initialArgs);
-};
-
-export const enhanceContext: LoaderFunction = (context) => {
-  if (globalThis.HTMLElement && context.canvasElement instanceof globalThis.HTMLElement) {
-    context.canvas = within(context.canvasElement);
-  }
-  context.userEvent = userEvent.setup();
-};
-
-export default () =>
-  definePreview({
-    loaders: [resetAllMocksLoader, nameSpiesAndWrapActionsInSpies, enhanceContext],
-  });
