+++ code/renderers/react/src/__test__/portable-stories.test.tsx
@@ -4,10 +4,9 @@ import { afterEach, beforeAll, describe, expect, it, vi } from 'vitest';
 
 import React from 'react';
 
-import type { ProjectAnnotations } from 'storybook/internal/csf';
 import { addons } from 'storybook/internal/preview-api';
 
-import type { Meta, ReactRenderer } from '@storybook/react';
+import type { Meta } from '@storybook/react';
 
 import { expectTypeOf } from 'expect-type';
